crawler/requirements.txt
# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# HTTP + rendering
playwright>=1.42.0
httpx[http2]>=0.27.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0
extruct>=0.17.0
w3lib>=2.1.2
trafilatura>=1.8.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Date parsing
python-dateutil>=2.9.0
pytz>=2024.1
dateparser>=1.2.0

# Reliability
tenacity>=8.2.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Database (direct Postgres / asyncpg — PgBouncer Transaction Pooler compatible)
asyncpg>=0.29.0

# ── LLM fallback providers (install only the one you need) ───────────────────
# Default provider — LLM_PROVIDER=gemini (new unified SDK, replaces google-generativeai)
google-genai>=1.0.0

# Required for semantic search embeddings (text-embedding-3-small)
# Also usable as LLM_PROVIDER=openai for event extraction
openai>=1.30.0

# Optional — LLM_PROVIDER=anthropic (also required for CHAT_PROVIDER=anthropic)
anthropic>=0.30.0

# Auth — Supabase JWT verification
PyJWT>=2.8.0

# Utilities
urllib3>=2.2.0

# File uploads — required by FastAPI for multipart/form-data (media upload endpoint)
python-multipart>=0.0.9
